refactor: schema to table
