+++ plugins/reflect/pkg/table/frame_test.go
@@ -1,4 +1,4 @@
-package schema
+package table
 
 import (
 	"context"
