+++ plugins/reflect/pkg/table/symbol.go
@@ -1,4 +1,4 @@
-package schema
+package table
 
 import (
 	"context"
